BUGS: initializeRemotely method. 1. Sometimes the remote database which we was getting elgamal    information(safe primes) from it goes offline and this makes our    module nothing but an useless code. 2. Also In some situations, we are unable to connect the server which    again causes problem. e.g. you can see this issue: https://github.com/nodejs/docker-node/issues/1653
NOTE: We call both of these situations, "Offline mode".

Solution:
Because of this problem, we decided to dump some of
information from remote database and store them locally. In result of
this decision, we provided 100 unique groups per 2048, 3072, 4096 and
8192 orders.
This solution is applied to both Node.js and browser but in browser
Since we don't have access to the file system, we always use default
group(first group of each order).

Overhead:
This solution brings 2MB overhead in disk space which I think it's
economical.

NOTE: This information is public and even we can use same group always.
It just matter of choice to have 100 group per each order.